Added code for ease of use and stability
Added addtional server arguments.
For people looking to run the AI in a standalone example you can do so by using the --force_train and --dir arguements. Instructions are given in the homepage readme.

Added additional error and warning messages to be more informative.

Added two java files to add to the base MicroRTS project to allow interaction with UTS_Imass
The socketAI file is a small modification to the base SocketAI .java but where the default port and language are set to the correct defaults for our server.

The FEStatePane.java was modified to import the UTS_Imass bot as the 2nd option in the drop down menu. It will also appear in the tournament tab