Bot stops executing when closing on time budget
Added a feature that will assign noop to units when exceeding 93% of the assigned action decision time. This is mainly to avoid timeout in tournament settings where computation may exceed the time limits for unknown reasons.

This doesn't effect the slave bots used for self learning

Additionally, added an argument (--ignore_budget) to ignore the timebudget requirement (Mostly used for debugging purposes).